rename geolocation field annotation to follow ESA convention, and move/rename burst methods from s1meta to sentinel1dataset with documentation changed
